+++ CHANGELOG.md
@@ -3,9 +3,11 @@
 TODO: Version 1.3
 --------------------------------------
 New Rules
+ - There is a new "comments" ruleset.
  - #359: **ClassEndsWithBlankLine** rule: Check whether the class ends with a blank line. (David Ausín)
  - #362: **ClassStartsWithBlankLine** rule: Check whether the class starts with a blank line. (David Ausín)
  - #377: **ExplicitCallToPutAtMethod** rule: Detects when the `map.putAt(k, v)` method is called directly rather than using `map[k] = v`.
+ - #385: **JavadocEmptyFirstLine** rule (comments) - Check for javadoc comments with an empty top line.
 
 Updated/Enhanced Rules and Bug Fixes
  - #364: **UnnecessarySetter** rule: Fix StringIndexOutOfBoundsException for single-letter property names (e.g. account.setE(3)).